Fix compiler error: conflicting types
../src/dyn_macosx.c:103:5: error: conflicting types for 'DYN_CloseLibrary'
  103 | int DYN_CloseLibrary(void *pvLHandle)
      |     ^
../src/dyn_generic.h:45:7: note: previous declaration is here
   45 |         LONG DYN_CloseLibrary(void *);
      |              ^
../src/dyn_macosx.c:119:5: error: conflicting types for 'DYN_GetAddress'
  119 | int DYN_GetAddress(void *pvLHandle, void **pvFHandle, const char *pcFunction,
      |     ^
../src/dyn_generic.h:46:7: note: previous declaration is here
   46 |         LONG DYN_GetAddress(void *, /*@out@*/ void **, const char *, bool);
      |              ^
../src/dyn_macosx.c:128:10: warning: implicit conversion from 'LONG' (aka 'long') to 'int' changes value from 2148532230 to -2146435066 [-Wconstant-conversion]
  128 |                 return SCARD_E_NO_MEMORY;
      |                 ~~~~~~ ^~~~~~~~~~~~~~~~~
./pcsclite.h:119:29: note: expanded from macro 'SCARD_E_NO_MEMORY'
  119 | #define SCARD_E_NO_MEMORY               ((LONG)0x80100006) /**< Not enough memory available to complete this command. */
      |                                          ^~~~~~~~~~~~~~~~
../src/dyn_macosx.c:133:10: warning: implicit conversion from 'LONG' (aka 'long') to 'int' changes value from 2148532244 to -2146435052 [-Wconstant-conversion]
  133 |                 return SCARD_F_UNKNOWN_ERROR;
      |                 ~~~~~~ ^~~~~~~~~~~~~~~~~~~~~
./pcsclite.h:147:33: note: expanded from macro 'SCARD_F_UNKNOWN_ERROR'
  147 | #define SCARD_F_UNKNOWN_ERROR           ((LONG)0x80100014) /**< An internal error has been detected, but the source is unknown. */
      |                                          ^~~~~~~~~~~~~~~~
2 warnings and 2 errors generated.

+++ src/dyn_macosx.c
@@ -94,7 +94,7 @@ void * DYN_LoadLibrary(const char *pcLibrary)
 	return pvLHandle;
 }
 
-int DYN_CloseLibrary(void *pvLHandle)
+LONG DYN_CloseLibrary(void *pvLHandle)
 {
 
 	CFBundleRef bundle = (CFBundleRef) pvLHandle;
@@ -110,7 +110,7 @@ int DYN_CloseLibrary(void *pvLHandle)
 	return SCARD_S_SUCCESS;
 }
 
-int DYN_GetAddress(void *pvLHandle, void **pvFHandle, const char *pcFunction,
+LONG DYN_GetAddress(void *pvLHandle, void **pvFHandle, const char *pcFunction,
 	bool mayfail)
 {
 	(void)mayfail;
